+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [24.0.0]
-- [BreakingChange][StateView] Added the ability for consumer to add a RefreshView to every state view.
+- [BreakingChange][StateView] Added the ability for consumer to add a RefreshView to some state views.
 
 ## [23.2.2]
 - Fixed an issue where title and description on EmptyView and ErrorView were not horizontally aligned